nasmlib.h: use int for the enum macros, for strict ISO C
ISO C before C23 restricted enum ranges to type "int" (not even
"unsigned int"). This limits the usability of enums for bit fields,
but it still covers our main use cases.

On gcc/clang this just produces a warning, but who knows what other
compilers will do.

+++ include/nasmlib.h
@@ -85,7 +85,15 @@ typedef union intorptr intorptr;
 # define fieldenc_cast(x,y) ((typeof(x ## _MASK))(y))
 # define fieldval_cast(x,y) ((typeof(x ## _BASEMASK))(y))
 #else
-# define fieldenc_cast(x,y) ((uintmax_t)(y))
+/*
+ * Using int here matches the strict ISO C before C23 which only allows
+ * an enum to be in the range of "int". This also means that these macros
+ * are only usable for bitfields up to 32 bits wide, which is extremely
+ * unfortunate, but covers most of all the NASM use cases.
+ *
+ * The (int) should at least give a warning on overflow.
+ */
+# define fieldenc_cast(x,y) ((int)(y))
 # define fieldval_cast(x,y) (y)
 #endif
